requirements.txt
appdirs==1.4.4
Authlib==0.15.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Deprecated==1.2.12
fastapi==0.68.1
uvicorn==0.15.0
github3.py==2.0.0
idna==2.10
jwcrypto==0.9.1
pycparser==2.20
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
simple-salesforce==1.11.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
xmltodict==0.12.0

